spark/openai/requirements.txt
spark-nlp==5.3.3
openai
python-dotenv